added functioning collision detection and angle adjustment based on where ball hits paddle

+++ src/Distance.java
@@ -3,8 +3,8 @@ public class Distance {
 
     }
     public static double getDistance(Position start, Position end) {
-        int deltaX = end.x - start.x;
-        int deltaY = end.y - start.y;
+        double deltaX = end.x - start.x;
+        double deltaY = end.y - start.y;
         return Math.sqrt(deltaX * deltaX + deltaY * deltaY);
     }
 }
