+++ src/Wall.java
@@ -9,8 +9,8 @@ public class Wall {
     }
     public Position getRandomPosition() {
         Random random = new Random();
-        int randomX = random.nextInt(start.x, end.x + 1);
-        int randomY = random.nextInt(start.y, end.y + 1);
+        int randomX = random.nextInt((int)start.x, (int)end.x + 1);
+        int randomY = random.nextInt((int)start.y, (int)end.y + 1);
         return new Position(randomX, randomY);
     }
 }
